+++ Packages/com.alelievr.mixture/Editor/Views/SelfNodeView.cs
@@ -15,12 +15,16 @@ namespace Mixture
 	{
 		SelfNode		selfNode;
 
-
 		public override void Enable(bool fromInspector)
 		{
 			base.Enable(fromInspector);
 			selfNode = nodeTarget as SelfNode;
 
+			controlsContainer.Add(new IMGUIContainer(() => {
+				// UIElements popup field are a pain to make
+				var elementNames = owner.graph.outputNode.outputTextureSettings.Select(t => t.name);
+				selfNode.outputIndex = EditorGUILayout.Popup("Output", selfNode.outputIndex, elementNames.ToArray());
+			}));
             controlsContainer.Add(new Button(ResetTexture) { text = "Reset"});
 		}