+++ CHANGELOG.md
@@ -10,6 +10,7 @@ https://user-images.githubusercontent.com/6877923/123006036-64e2e780-d3b7-11eb-9
 
 - Added conversion methods for Vector3 and Vector2 to Vector4.
 - Added a popup to select which preview texture is visible in the output node.
+- Added an option in the self node to select the output texture.
 
 ### Fixed
 - Fixed missing nodes in the list of nodes proposed when dragging an edge.